+++ app/src/main/java/com/github/siela1915/bootcamp/UploadRecipe/ReviewRecipeBeforeUploadingDialog.java
@@ -1,4 +1,4 @@
-package com.github.siela1915.bootcamp;
+package com.github.siela1915.bootcamp.UploadRecipe;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -13,6 +13,8 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.fragment.app.DialogFragment;
+
+import com.github.siela1915.bootcamp.R;
 
 import java.io.IOException;
