+++ app/src/main/java/com/github/siela1915/bootcamp/UploadRecipe/TextValidator.java
@@ -1,4 +1,4 @@
-package com.github.siela1915.bootcamp;
+package com.github.siela1915.bootcamp.UploadRecipe;
 
 import android.text.Editable;
 import android.text.TextWatcher;
